Update README.pl.md
In line 61 
    “imienia” should be replaced by “nazwiska”
Name is translated in Polish  to “imię” ( appropriate form in this instance is “imienia”) only in the very direct/strict translation and means first name. In this case more appropriate will be “nazwiska” (a form of nazwisko) which in the very direct/strict translation means last name, but is commonly understood as first name and last name given together.