Drawing more attention to 'Additional Info'
Rather than a somewhat out-of-place H3 referencing "Additional Information", it seemed benign to flesh the reference out in to a paragraph format comparable with the rest of the "next steps" section, in order to draw more attention to the content therein. Several issues  seem to  request touching on advanced techniques which are already outlined comprehensively under the less-visible Additional Information directory, and this change may make that more readily apparent.