fix: access static flag icons from the up-to-date repository
While ago, the original Github user name of the 'country-flags' repository has being changed:

 - Old: https://github.com/hjnilsson/country-flags
 + New: https://github.com/hampusborgos/country-flags

Despite the fact that Github has request routing in-place to keep old URLs working it forces web browsers to do additional HTTP redirect requests to access the final resource. Taking into account the high number of resources (flags) the main page is referring to it makes sense to switch to the new URL and safe some time on the unnecessary network routing.