Update README.md
On line 170. I have added a line to describing the action item further where the learner has to click to make a successful PR.

Note: I just completed my first-contributions tutorial and was stuck at line 170 for few mins. Hence, wanted to add a description of where the learner must click to complete the PR