README.kz.md word mistranslation update
Changed the word "коммандалық" to a better fitting word "топтық". "Kоммандалық" is rarely used for describing team/group projects, it's more of a Russian loanword used for describing sports team, debates team or anything alike. "Tоптық" translates as a group of people, and is more often used than the previous word, it is a direct translation from the English and not from Russian which is probably the reason "коммандалық" was used instead of my suggestion. I believe it would be more comfortable for Kazakh speakers to understand it and not get confused because of the meaning of the word.

This is my first ever contribution to the open source and I hope I am not breaking rules or confusing anyone, I am open to any discussions if needed.