Change correct Yoruba flag
Incorrect Flag for Nigerian Languages:
Multiple languages from Nigeria (Igbo, Yoruba, and Hausa) are using the same flag but different URLs. This is fine if intentional but can be confusing.

Fix:
Ensure that the flags and URLs are correctly mapped to the respective translations. Here’s an for Yoruba: https://www.fotw.info/images/n/ng%7Deoyor.gif