Update README.md
The example command for git clone is written using SSH, but the image next to the code shows the HTTPS option. This inconsistency may confuse users.

MAKE itclearer so that users can more easily understand the process of contributing. Ensuring that the image matches the example command will help users follow the steps correctly.

So, I replaced the SSH command with the HTTPS command and enhanced the bottom of the document.