Update README.md
Hey together!
As part of a university project in the module "Open Source Software," I have been tasked with identifying opportunities for improvement in open-source projects. 

Proposed Change:
Include a short section in the introduction of the README that briefly explains what a GitHub contribution is. This could help orient new contributors and make the project more beginner-friendly.

Best regards
domnomg