Spanish flag name (#94665)
Spanish flag name is a bit weird. 'Española' is an adjetive to refer to
something that is spanish like the language or the flag.
Instead 'Español' is the sustantive to refer to the languague,
just like they have use for other countries in the project.

I added the whole 'Español de España' to be consistent with the
flags from México and Ecuador. I am sure more latin contributors
will come soon and they will add more flags to the project.